requirements.txt
# requirements.txt
pydantic==2.9.2
python-dotenv==1.0.1
rich==13.9.4
loguru==0.7.2
requests==2.32.3
httpx==0.27.2
tenacity==9.0.0
tqdm==4.66.5
numpy==2.1.3
pandas==2.2.3
pytest==8.3.3
